ProjectProvider - Use the time id as the id for the time child.
Modified what the time child id will be, previously the position + offset
was used which works but is not logical (since each item has a guaranteed
unique id).

Also, using the actual id will help search for a specific item.

+++ app/src/main/java/me/raatiniemi/worker/util/TimesheetExpandableDataProvider.java
@@ -86,16 +86,16 @@ public class TimesheetExpandableDataProvider {
     }
 
     public static abstract class Data<T> {
-        private int mId;
+        private long mId;
 
         private T mData;
 
-        public Data(int id, T data) {
+        public Data(long id, T data) {
             mId = id;
             mData = data;
         }
 
-        public int getId() {
+        public long getId() {
             return mId;
         }
 
@@ -105,11 +105,11 @@ public class TimesheetExpandableDataProvider {
     }
 
     public static class TimeGroup extends Data<Date> {
-        public TimeGroup(int id, Date date) {
+        public TimeGroup(long id, Date date) {
             super(id, date);
         }
 
-        public int getGroupId() {
+        public long getGroupId() {
             return getId();
         }
 
@@ -119,11 +119,11 @@ public class TimesheetExpandableDataProvider {
     }
 
     public static class TimeChild extends Data<Time> {
-        public TimeChild(int id, Time time) {
+        public TimeChild(long id, Time time) {
             super(id, time);
         }
 
-        public int getChildId() {
+        public long getChildId() {
             return getId();
         }
